fix: make shims `cargo test --workspace` compatible
